+++ CHANGELOG.md
@@ -1,8 +1,22 @@
+## 1.123.0
+
+- adding in signin flow pages
+- adding in finalize account pages
+- adding in start of basic admin page
+- adding in start of common header
+- fixing sticky note error
+- updating fonts, theme, and gloabl css files
+- add in new logo files
+
 ## 1.122.0
 
-- Only send device id options in useLocalVideoToggle > getLocalVideoTrack if the device id ref has a truthy value.
-- This corrects the behavior inside the room of the browser asking permission to access a camera if one had not been enabled yet.
-- Passing a falsey-yet-defined value for the device id option to the video track was resulting in an overconstrained error, which propagated up as a permission error and the browser not asking for camera permissions when inside the room.
+- adding in signin flow pages
+- adding in finalize account pages
+- adding in start of basic admin page
+- adding in start of common header
+- fixing sticky note error
+- updating fonts, theme, and gloabl css files
+- add in new logo files
 
 ## 1.121.0